Extract popup quick actions from content script
